+++ design-toolkit/components/src/components/classification-banner/index.tsx
@@ -10,16 +10,14 @@
  * governing permissions and limitations under the License.
  */
 
+import type { ProviderProps } from '@/lib/types';
 import { createContext } from 'react';
 import { type ContextValue, useContextProps } from 'react-aria-components';
 import {
   ClassificationBannerStyles,
   ClassificationBannerStylesDefaults,
 } from './styles';
-import type {
-  ClassificationBannerProps,
-  ClassificationBannerProviderProps,
-} from './types';
+import type { ClassificationBannerProps } from './types';
 
 export const ClassificationBannerContext =
   createContext<ContextValue<ClassificationBannerProps, HTMLDivElement>>(null);
@@ -27,7 +25,7 @@ export const ClassificationBannerContext =
 function ClassificationBannerProvider({
   children,
   ...props
-}: ClassificationBannerProviderProps) {
+}: ProviderProps<ClassificationBannerProps>) {
   return (
     <ClassificationBannerContext.Provider value={props}>
       {children}