+++ design-toolkit/components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @accelint/design-toolkit
 
+## 2.4.3
+
+### Patch Changes
+
+- 382bbc2: Fixes a build issue caused by an import with trailing character in `components/avatar/index.tsx`.
+
+## 2.4.2
+
+### Patch Changes
+
+- 93b3750: Fixed controlled usage of field input
+
 ## 2.4.1
 
 ### Patch Changes